Do not use uint8_t for the number of operand in VASTExpr, since the underlying  number of operand in the OperandList class have an unsigned type.
patched by yangliwei.uestc@gmail.com.

+++ include/shang/VASTDatapathNodes.h
@@ -180,7 +180,7 @@ private:
   VASTExpr(const VASTExpr&);              // Do not implement
   void operator=(const VASTExpr&);        // Do not implement
 
-  VASTExpr(Opcode Opc, uint8_t NumOps, unsigned UB, unsigned LB);
+  VASTExpr(Opcode Opc, unsigned NumOps, unsigned UB, unsigned LB);
   VASTExpr();
 
   friend struct ilist_sentinel_traits<VASTExpr>;
